refactor(ui): move lib/beak-variable-set → services/variable-sets (ADR 0004)
Migrates packages/ui/src/lib/beak-variable-set/ (singular) to
packages/ui/src/services/variable-sets/ (plural), matching the slice
name in @beak/state/variable-sets.

Updated the relative imports inside the moved index.ts (../fs, ../ipc →
../../lib/fs, ../../lib/ipc) and retargeted the three consumer import
paths: store/effects/variable-sets.ts, features/variables/parser.ts, and
features/variable-sets/components/molecules/VariableCard.tsx.

No function names, signatures, or behaviour changed — path migration only.
