+++ packages/ui/src/store/effects/variable-sets.ts
@@ -2,7 +2,7 @@ import { TypedObject } from '@beak/common/helpers/typescript';
 import { attemptReconciliation, changeTab } from '@beak/ui/features/tabs/store/actions';
 import type { ActiveRename } from '@beak/ui/features/tree-view/types';
 import { createVariableSet, renameVariableSet } from '@beak/ui/lib/beak-project/variable-sets';
-import { readVariableSet, removeVariableSet, writeVariableSet } from '@beak/ui/lib/beak-variable-set';
+import { readVariableSet, removeVariableSet, writeVariableSet } from '@beak/ui/services/variable-sets';
 import createFsEmitter, { type FsSubscription, scanDirectoryRecursively } from '@beak/ui/lib/fs-emitter';
 import { ipcDialogService, ipcFsService } from '@beak/ui/lib/ipc';
 import type { VariableSet, VariableSets } from '@getbeak/types/variable-sets';
